TASK: Correct test command and wtf?
