+++ composer.json
@@ -25,7 +25,7 @@
         "lint": "composer run-script check-psr2 && composer run-script self-check",
         "check-psr2": "phpcs --standard=PSR2 src/**",
         "self-check": "bin/editorconfig-checker -d -e 'vendor|.git|.png|.lock' ./*",
-        "test": "phpunit tests --bootstrap vendor/autoload.php"
+        "test": "phpunit tests/**/*.php --bootstrap vendor/autoload.php"
     },
     "bin": [
         "bin/editorconfig-checker"
